refactor: rename cypress elements to camelCase
