+++ cypress/elements/confirmLeaveModal.js
