+++ cypress/elements/dimensionModal/periodDimension.js
