+++ cypress/elements/dimensionsPanel.js
