+++ cypress/elements/optionsModal/fontStyles.js
