+++ cypress/elements/optionsModal/index.js
