+++ cypress/elements/optionsModal/lines.js
