+++ cypress/elements/optionsModal/verticalAxis.js
