+++ cypress/elements/startScreen.js
