+++ cypress/elements/visualizationTypeSelector.js
