+++ cypress/elements/window.js
