+++ cypress/integration/options/verticalAxis.cy.js
@@ -1,14 +1,14 @@
 import { DIMENSION_ID_DATA, VIS_TYPE_COLUMN } from '@dhis2/analytics'
 
-import { openDimension } from '../../elements/DimensionsPanel'
+import { openDimension } from '../../elements/dimensionsPanel'
 import {
     selectDataElements,
     clickDimensionModalUpdateButton,
-} from '../../elements/DimensionModal'
-import { goToStartPage } from '../../elements/StartScreen'
-import { expectVisualizationToBeVisible } from '../../elements/Chart'
+} from '../../elements/dimensionModal'
+import { goToStartPage } from '../../elements/startScreen'
+import { expectVisualizationToBeVisible } from '../../elements/chart'
 import { TEST_DATA_ELEMENTS } from '../../utils/data'
-import { clickMenuBarOptionsButton } from '../../elements/MenuBar'
+import { clickMenuBarOptionsButton } from '../../elements/menuBar'
 import {
     clickOptionsModalUpdateButton,
     clickOptionsTab,
@@ -20,7 +20,7 @@ import {
     setVerticalAxisRangeMaxValue,
     setVerticalAxisRangeMinValue,
     setVerticalAxisTitle,
-} from '../../elements/OptionsModal'
+} from '../../elements/optionsModal'
 import {
     expectStoreConfigYAxisToHaveRangeMinValue,
     expectStoreConfigYAxisToHaveRangeMaxValue,
